requirements.txt
requests==2.31.0
pytube==15.0.0
pygame==2.5.0
pydub==0.25.1
python-dotenv==1.0.0